requirements.txt
algoliasearch
bcrypt
millify
naff
orjson
pillow
pygithub
PyMySQL
python-dotenv
samp-client
